+++ src/js/mixin/position.js
@@ -3,6 +3,7 @@ import {
     flipPosition,
     getCssVar,
     offset as getOffset,
+    includes,
     isNumeric,
     isRtl,
     positionAt,
@@ -44,25 +45,57 @@ export default {
             }
             offset += toPx(getCssVar('position-margin-offset', element));
 
-            const { x, y } = positionAt(
-                element,
-                target,
-                axis === 'x' ? `${flipPosition(dir)} ${align}` : `${align} ${flipPosition(dir)}`,
-                axis === 'x' ? `${dir} ${align}` : `${align} ${dir}`,
-                axis === 'x'
-                    ? `${dir === 'left' ? -offset : offset}`
-                    : ` ${dir === 'top' ? -offset : offset}`,
-                null,
-                this.flip,
-                boundary
-            ).target;
+            positionAt(element, target, {
+                boundary,
+                flip: this.flip,
+                attach: {
+                    element: axis === 'x' ? [flipPosition(dir), align] : [align, flipPosition(dir)],
+                    target: axis === 'x' ? [dir, align] : [align, dir],
+                },
+                offset:
+                    axis === 'x'
+                        ? [dir === 'left' ? -offset : +offset, 0]
+                        : [0, dir === 'top' ? -offset : +offset],
+            });
 
-            this.dir = axis === 'x' ? x : y;
-            this.align = axis === 'x' ? y : x;
+            [this.dir, this.align] = getAlignment(element, target);
         },
 
         getAxis() {
-            return this.dir === 'top' || this.dir === 'bottom' ? 'y' : 'x';
+            return includes(['top', 'bottom'], this.dir) ? 'y' : 'x';
         },
     },
 };
+
+function getAlignment(el, target) {
+    const elOffset = getOffset(el);
+    const targetOffset = getOffset(target);
+    const properties = [
+        ['left', 'right'],
+        ['top', 'bottom'],
+    ];
+
+    let dir;
+    for (const props of properties) {
+        if (elOffset[props[0]] > targetOffset[props[1]]) {
+            dir = props[1];
+            break;
+        }
+        if (elOffset[props[1]] < targetOffset[props[0]]) {
+            dir = props[0];
+            break;
+        }
+    }
+
+    let align;
+    const props = includes(properties[0], dir) ? properties[1] : properties[0];
+    if (elOffset[props[0]] === targetOffset[props[0]]) {
+        align = props[0];
+    } else if (elOffset[props[1]] === targetOffset[props[1]]) {
+        align = props[1];
+    } else {
+        align = 'center';
+    }
+
+    return [dir, align];
+}