+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## WIP
 
+### Added
+
+- Drop/Dropdowns try to shift to stay within viewport
+
 ### Changed
 
 - Use custom property for direction offset in components based on Position mixin