+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Fix Switcher component with disabled toggles
 - Fix responsiveness in Responsive component
 - Fix video autoplay
+- Fix importing UIkit in Next.js
 
 ## 3.13.1 (March 18, 2022)